+++ src/dominion/models/persona/UnknownPersona.java
@@ -3,7 +3,7 @@ package dominion.models.persona;
 import dominion.models.persona.name.PersonaName;
 import dominion.models.persona.name.UnknownPersonaName;
 
-public class UnknowPersona implements Persona{
+public class UnknownPersona implements Persona{
 
     private PersonaName name = new UnknownPersonaName();